Added first Project referee - WorkflowGraph.
